requirements.txt
Flask==2.3.3
PyPDF2==3.0.1
pdfplumber==0.10.3
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
spacy==3.6.1
python-docx==1.1.0
pillow==10.0.0
gunicorn==21.2.0
pytesseract==0.3.10
pdf2image==1.16.3
PyMuPDF==1.23.4